src/main/java/io/blog/springblogapp/service/impl/UserDetailsServiceImpl.java
package io.blog.springblogapp.service.impl;

import io.blog.springblogapp.model.entity.UserEntity;
import io.blog.springblogapp.repository.UserRepository;
import lombok.AllArgsConstructor;
import org.springframework.context.annotation.Configuration;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;

import java.util.Optional;

@AllArgsConstructor
@Configuration
public class UserDetailsServiceImpl implements UserDetailsService {

    private final UserRepository userRepository;

    @Override
    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
        Optional<UserEntity> user = userRepository.findByEmail(email);
        if (user.isEmpty()) throw new UsernameNotFoundException("Wrong credentials.");

        return user.get();
    }
}
